+++ src/main/java/kr/pe/timeorder/controller/StoreController.java
@@ -5,6 +5,7 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
+import org.hibernate.criterion.NotExpression;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,8 +17,14 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import kr.pe.timeorder.exception.InvalidException;
+import kr.pe.timeorder.exception.NotFoundException;
+import kr.pe.timeorder.exception.PermissionException;
+import kr.pe.timeorder.exception.TokenException;
+import kr.pe.timeorder.model.Item;
 import kr.pe.timeorder.model.Member;
 import kr.pe.timeorder.model.Store;
+import kr.pe.timeorder.repository.ItemRepository;
 import kr.pe.timeorder.repository.MemberRepository;
 import kr.pe.timeorder.repository.StoreRepository;
 import kr.pe.timeorder.service.JwtService;
@@ -31,7 +38,10 @@ public class StoreController {
 	
 	@Autowired
 	private MemberRepository mRepository;
-
+	
+	@Autowired
+	private ItemRepository iRepository;
+	
 	@Autowired
 	private JwtService jwtService;
 	
@@ -41,15 +51,22 @@ public class StoreController {
 		HttpStatus status = null;
 		List<Store> list = null;
 		try {
-			jwtService.checkValid(req.getHeader("jwt-auth-token"));
-			LinkedHashMap l = (LinkedHashMap)(jwtService.get(req.getHeader("jwt-auth-token")).get("Member"));
-			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"));
-			if (loginMember.getAuthor() != 2) {
-				throw new RuntimeException();
+			String token = req.getHeader("jwt-auth-token");
+			if (token == null || token.length() == 0) {
+				throw new TokenException();
 			}
-			System.out.println(loginMember);
-			status = HttpStatus.ACCEPTED;
+			jwtService.checkValid(token);
+			LinkedHashMap l = (LinkedHashMap)(jwtService.get(token).get("Member"));
+			
+			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"))
+					.orElseThrow(()-> new NotFoundException("member"));
+			
+			if (loginMember.getAuthor() != 2) {
+				throw new PermissionException();
+			}
+			
 			list = sRepository.findAll();
+			status = HttpStatus.ACCEPTED;
 		} catch(RuntimeException e) {
 			log.error("정보 조회 실패 ", e);
 			status = HttpStatus.INTERNAL_SERVER_ERROR;
@@ -62,20 +79,22 @@ public class StoreController {
 		HttpStatus status = null;
 		List<Store> list = null;
 		try {
-			jwtService.checkValid(req.getHeader("jwt-auth-token"));
-			LinkedHashMap l = (LinkedHashMap)(jwtService.get(req.getHeader("jwt-auth-token")).get("Member"));
-			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"));
+			String token = req.getHeader("jwt-auth-token");
+			if (token == null || token.length() == 0) {
+				throw new TokenException();
+			}
+			jwtService.checkValid(token);
+			LinkedHashMap l = (LinkedHashMap)(jwtService.get(token).get("Member"));
+			
+			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"))
+					.orElseThrow(()-> new NotFoundException("member"));
 			Member findMember = mRepository.findById(loginMemberId).get();
 			
-			if (findMember.getAuthor() != 1) {
-				throw new RuntimeException();
-			}
-			
-			if (loginMember.getAuthor() == 2 || loginMember.getMemberId() == loginMemberId) {
-				status = HttpStatus.ACCEPTED;
+			if (loginMember.getAuthor() > 0 || loginMember.getMemberId() == loginMemberId) {
 				list = sRepository.findStoreByMember(findMember);
+				status = HttpStatus.ACCEPTED;
 			} else {
-				throw new RuntimeException();
+				throw new PermissionException();
 			}
 		} catch(RuntimeException e) {
 			log.error("정보 조회 실패 ", e);
@@ -83,22 +102,50 @@ public class StoreController {
 		}
 		return new ResponseEntity<List<Store>>(list, status);
 	}
+	
+	@GetMapping("/stores/item/{itemid}")
+	public ResponseEntity<Store> itemStores(HttpServletRequest req, @PathVariable long itemid) {
+		HttpStatus status = null;
+		Item item = null;
+		try {
+			item = iRepository.findById(itemid).orElseThrow(() -> new NotFoundException());
+			if (item.getStore() == null) {
+				throw new NotFoundException();
+			}
+			status = HttpStatus.ACCEPTED;
+		} catch(RuntimeException e) {
+			log.error("정보 조회 실패 ", e);
+			status = HttpStatus.INTERNAL_SERVER_ERROR;
+		}
+		return new ResponseEntity<Store>(item.getStore(), status);
+	}
 
 	@PostMapping("/stores")
 	public ResponseEntity<Store> newStore(HttpServletRequest req, @RequestBody Store newStore) {
 		HttpStatus status = null;
 		try {
-			jwtService.checkValid(req.getHeader("jwt-auth-token"));
-			LinkedHashMap l = (LinkedHashMap)(jwtService.get(req.getHeader("jwt-auth-token")).get("Member"));
-			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"));
+			String token = req.getHeader("jwt-auth-token");
+			if (token == null || token.length() == 0) {
+				throw new TokenException();
+			}
+			jwtService.checkValid(token);
+			LinkedHashMap l = (LinkedHashMap)(jwtService.get(token).get("Member"));
+			
+			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"))
+					.orElseThrow(()-> new NotFoundException("mission"));
 			
 			if (loginMember.getAuthor() != 1) {
-				throw new RuntimeException();
+				throw new PermissionException();
 			}
 			
-			status = HttpStatus.ACCEPTED;
 			newStore.setMember(loginMember);
+
+			if (!newStore.isVaild()) {
+				throw new InvalidException();
+			}
+			
 			sRepository.save(newStore);
+			status = HttpStatus.ACCEPTED;
 		} catch(RuntimeException e) {
 			log.error("정보 조회 실패 ", e);
 			status = HttpStatus.INTERNAL_SERVER_ERROR;
@@ -114,24 +161,32 @@ public class StoreController {
 	@PutMapping("/stores/{storeId}")
 	public ResponseEntity<Store> replaceStore(HttpServletRequest req ,@RequestBody Store newStore, @PathVariable long storeId) {
 		HttpStatus status = null;
-		Store store = null;
+		Store store = null; 
 		try {
-			jwtService.checkValid(req.getHeader("jwt-auth-token"));
-			LinkedHashMap l = (LinkedHashMap)(jwtService.get(req.getHeader("jwt-auth-token")).get("Member"));
-			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"));
-			store = sRepository.findById(storeId).get();
-			if (loginMember.getAuthor() != 1 && loginMember.getMemberId() != store.getMember().getMemberId()) {
-				throw new RuntimeException();
+			String token = req.getHeader("jwt-auth-token");
+			if (token == null || token.length() == 0) {
+				throw new TokenException();
 			}
-			store.setItems(newStore.getItems());
-			store.setMember(newStore.getMember());
-			store.setReviews(newStore.getReviews());
+			jwtService.checkValid(token);
+			LinkedHashMap l = (LinkedHashMap)(jwtService.get(token).get("Member"));
+			
+			Member loginMember = mRepository.findMemberByPhone((String)l.get("phone"))
+					.orElseThrow(()-> new NotFoundException("member"));
+			store = sRepository.findById(storeId).get();
+			
+			if (loginMember.getAuthor() != 1 && loginMember.getMemberId() != store.getMember().getMemberId()) {
+				throw new PermissionException();
+			}
+			
 			store.setStoreInfo(newStore.getStoreInfo());
 			store.setStoreName(newStore.getStoreName());
 			store.setStoreNum(newStore.getStoreNum());
+			store.setCloseTime(newStore.getCloseTime());
+			store.setOpenTime(newStore.getOpenTime());
+			store.setAddress_name(newStore.getAddress_name());
+			store.setDetail_address(newStore.getDetail_address());
 			sRepository.save(store);
 			status = HttpStatus.ACCEPTED;
-			
 		} catch(RuntimeException e) {
 			log.error("정보 조회 실패 ", e);
 			status = HttpStatus.INTERNAL_SERVER_ERROR;
@@ -141,21 +196,28 @@ public class StoreController {
 
 	@DeleteMapping("/stores/{storeId}")
 	public ResponseEntity<Store> deleteStore(HttpServletRequest req, @PathVariable long storeId) {
-		log.info("---- updateEmployees () -----------------");
+		log.info("---- deleteStore () -----------------");
 		HttpStatus status = null;
 		Member loginMember = null;
 		Store store = null;
 		
 		try {
-			jwtService.checkValid(req.getHeader("jwt-auth-token"));
-			LinkedHashMap l = (LinkedHashMap)(jwtService.get(req.getHeader("jwt-auth-token")).get("Member"));
-			loginMember = mRepository.findMemberByPhone((String)l.get("phone"));
+			String token = req.getHeader("jwt-auth-token");
+			if (token == null || token.length() == 0) {
+				throw new TokenException();
+			}
+			jwtService.checkValid(token);
+			LinkedHashMap l = (LinkedHashMap)(jwtService.get(token).get("Member"));
+			
+			loginMember = mRepository.findMemberByPhone((String)l.get("phone"))
+					.orElseThrow(()-> new NotFoundException("member"));
 			store = sRepository.findById(storeId).get();
+			
 			if (loginMember.getAuthor() > 0 || loginMember.getMemberId() == store.getMember().getMemberId()) {
 				sRepository.deleteById(storeId);
 				status = HttpStatus.ACCEPTED;
 			} else {
-				throw new RuntimeException();
+				throw new PermissionException();
 			}
 		} catch(RuntimeException e) {
 			log.error("정보 조회 실패 ", e);
